Escape SSID and password in POST request
If SSID or password contain any weird characters they may not be stored
properly. Escape using `encodeURIComponent()` to make sure they are
formatted properly before.

A better way to solve this would be to use `fetch()` with `FormData`
which is the modern way to do HTTP requests but this requires some more
work since the entire request function would be need to be rewritten.